Continued kickoff sequence, using the System class for time management will have to be avoided though

+++ src/main/java/net/akami/mistream/play/list/LinearEndToEndSequence.java
@@ -15,7 +15,7 @@ public class LinearEndToEndSequence extends EndToEndSequence {
     }
 
     @Override
-    protected Function<Integer, Float> getBoostFunction() {
+    protected Function<Long, Float> getBoostFunction() {
         return (i) -> speed;
     }
 }
